Renaming 'webrenovation' to 'usurper'
