+++ spec/usurper/pages/az_databases_page.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-module WebRenovation
+module Usurper
   module Pages
     # /personal
     class AZDatabases < BasePage
