+++ spec/usurper/pages/az_subjects_page.rb
@@ -1,4 +1,4 @@
-module WebRenovation
+module Usurper
   module Pages
     class AZSubjects < BasePage
       include Capybara::DSL
