+++ spec/usurper/pages/library_giving_page.rb
@@ -1,20 +1,24 @@
 # frozen_string_literal: true
-module WebRenovation
+module Usurper
   module Pages
     # /personal
-    class LibraryPages < BasePage
+    class LibraryGivingPage
       include Capybara::DSL
       include CapybaraErrorIntel::DSL
 
       def on_page?
-        super &&
+        correct_content? &&
         on_valid_url?
+      end
+
+      def correct_content?
+        page.has_link?('Library Giving')
       end
 
       def on_valid_url?
         last_opened_window = page.driver.browser.window_handles.last
         page.driver.browser.switch_to_window(last_opened_window)
-        current_url.include?('https://www.google.com/maps/place/')
+        current_url == ('http://librarygiving.nd.edu/')
       end
     end
   end
