+++ spec/usurper/pages/library_jobs_page.rb
@@ -1,24 +1,24 @@
 # frozen_string_literal: true
-module WebRenovation
+module Usurper
   module Pages
     # /personal
-    class LibraryGivingPage
+    class LibraryJobsPage < BasePage
       include Capybara::DSL
       include CapybaraErrorIntel::DSL
 
       def on_page?
+        super &&
         correct_content? &&
         on_valid_url?
       end
 
       def correct_content?
-        page.has_link?('Library Giving')
+        page.has_css?('h2', text: 'Employment')
+        page.has_link?('Office of Human Resources, Employment Opportunities')
       end
 
       def on_valid_url?
-        last_opened_window = page.driver.browser.window_handles.last
-        page.driver.browser.switch_to_window(last_opened_window)
-        current_url == ('http://librarygiving.nd.edu/')
+        current_url == File.join(Capybara.app_host, "employment/")
       end
     end
   end
