+++ spec/usurper/pages/reseach_guides_page.rb
@@ -1,5 +1,5 @@
 
-module WebRenovation
+module Usurper
   module Pages
     class ResearchGuidesPage < BasePage
       include Capybara::DSL
