+++ spec/usurper/pages/reserve_a_room_page_services_tab.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-module WebRenovation
+module Usurper
   module Pages
     # /personal
     class RoomReservationServicesTabPage < BasePage
