+++ spec/usurper/pages/search_page.rb
@@ -1,4 +1,4 @@
-module WebRenovation
+module Usurper
   module Pages
     class SearchPage
       include Capybara::DSL
